+++ src/kof_parser/writer.py
@@ -1,37 +1,38 @@
-from typing import List
+from typing import List, Optional
 from datetime import datetime
 from uuid import UUID
 
-from ngi_kof_parser import Kof
-from ngi_kof_parser import Location
-from ngi_kof_parser import MethodTypeEnum
+from kof_parser import Kof
+from kof_parser import Location
+from kof_parser.enums import MethodType
+from kof_parser import projector
 
 
 class KOFWriter(Kof):
 
     method_type_to_temakode = {
-        MethodTypeEnum.RO.name: "2251",
-        MethodTypeEnum.RWS.name: "2401",
-        MethodTypeEnum.SA.name: "2402",
-        MethodTypeEnum.TP.name: "2403",
-        MethodTypeEnum.SS.name: "2405",
-        MethodTypeEnum.RP.name: "2406",
-        MethodTypeEnum.CPT.name: "2407",
-        MethodTypeEnum.RS.name: "2409",
-        MethodTypeEnum.SR.name: "2410",
-        MethodTypeEnum.SPT.name: "2411",
-        MethodTypeEnum.RCD.name: "2412",
-        MethodTypeEnum.PZ.name: "2413",
-        MethodTypeEnum.PT.name: "2414",
-        MethodTypeEnum.SVT.name: "2415",
-        MethodTypeEnum.INC.name: "2417",
-        MethodTypeEnum.TOT.name: "2418",
+        MethodType.RO.name: "2251",
+        MethodType.RWS.name: "2401",
+        MethodType.SA.name: "2402",
+        MethodType.TP.name: "2403",
+        MethodType.SS.name: "2405",
+        MethodType.RP.name: "2406",
+        MethodType.CPT.name: "2407",
+        MethodType.RS.name: "2409",
+        MethodType.SR.name: "2410",
+        MethodType.SPT.name: "2411",
+        MethodType.RCD.name: "2412",
+        MethodType.PZ.name: "2413",
+        MethodType.PT.name: "2414",
+        MethodType.SVT.name: "2415",
+        MethodType.INC.name: "2417",
+        MethodType.TOT.name: "2418",
     }
 
     def __init__(self):
         super().__init__()
 
-    def create_admin_block(self, project_name: str, srid: int) -> str:
+    def create_admin_block(self, project_name: str, srid: int, swap_easting_northing: Optional[bool] = False) -> str:
         """
         # 00 Oppdrag Dato Ver K.sys Komm $11100000000 Observatør
         # 01 EXP          31012022   2      22 0000 $22100000000           NN
@@ -40,7 +41,10 @@ class KOFWriter(Kof):
         version = "1"
         code = self.get_code(srid=srid)
         municipality = ""
-        units = "$21100000000"
+        if swap_easting_northing:
+            units = "$21100000000"
+        else:
+            units = "$11100000000"
         observer = ""
 
         #             "-01 OOOOOOOOOOOO DDMMYYYY VVV KKKKKKK KKKK $RVAllllllll OOOOOOOOOOOO"
@@ -53,7 +57,7 @@ class KOFWriter(Kof):
 
     @staticmethod
     def create_kof_coordinate_block(id: str, temakode: str, x: float, y: float, z: float) -> str:
-        coord_block = f" 05 {id[0:10]:<10} {temakode:<8} {x:<12.3f} {y:<11.3f} {z:<8.3f} "
+        coord_block = f" 05 {id[0:10]:<10} {temakode:<8} {y:<12.3f} {x:<11.3f} {z:<8.3f} "
         coord_block = f"{coord_block:<70}\n"
         return coord_block
 
@@ -72,21 +76,27 @@ class KOFWriter(Kof):
         project_name: str,
         locations: List[Location],
         srid: int,
+        swap_easting_northing: Optional[bool] = False,
     ) -> str:
         """For now, we do not do any transformation of locations positions"""
 
         kof_string = self.create_kof_header_lines(project_id=project_id, project_name=project_name, srid=srid)
-        kof_string += self.create_admin_block(project_name, srid=srid)
-        for loc in locations:
+        kof_string += self.create_admin_block(project_name, srid=srid, swap_easting_northing=swap_easting_northing)
+        for location in locations:
+            location.name = location.name if not None else ""
+            z = location.point_z or 0
+            x = location.point_easting or 0
+            y = location.point_northing or 0
+            if x and y and srid and location.srid and srid != location.srid:
+                x, y = projector.transform(from_srid=location.srid, to_srid=srid, east=x, north=y)
 
-            z = loc.point_z or 0
-            x = loc.point_easting or 0
-            y = loc.point_northing or 0
+            if swap_easting_northing:
+                x, y = y, x
 
-            if len(loc.methods) > 0:
-                for method in loc.methods:
+            if location.methods:
+                for method in location.methods:
                     kof_string += self.create_kof_coordinate_block(
-                        id=loc.name,
+                        id=location.name,
                         temakode=self.method_type_to_temakode.get(method, ""),
                         x=x,
                         y=y,
@@ -94,7 +104,7 @@ class KOFWriter(Kof):
                     )
             else:
                 kof_string += self.create_kof_coordinate_block(
-                    id=loc.name,
+                    id=location.name,
                     temakode="",
                     x=x,
                     y=y,