+++ src/ngi_kof_parser/__init__.py
@@ -1,4 +1,0 @@
-from ngi_kof_parser.kof import *
-from ngi_kof_parser.model import *
-from ngi_kof_parser.parser import *
-from ngi_kof_parser.writer import *